LHA encoder: lh5 (LZSS + length-limited canonical Huffman) + lh0 writer
The encode-side counterpart of LHAArchive: level-0 headers, per-block
canonical Huffman mirroring LHDecoder's bitstream exactly (the decoder
is the format spec), package-merge length limiting, degenerate
single-symbol trees in the 0-bit count==0 form, stored fallback when
compression does not pay, deterministic archives via a fixed epoch.
CLI: 'lha create <out> <dir> [--store] [--epoch N]' + 'lha extract'.
7 round-trip tests incl. system-lha (lhasa) CRC interop.

+++ Sources/AmigaDiskKit/ImageIO/BlockDevice.swift
@@ -11,7 +11,7 @@ public struct HostFileIdentity: Equatable {
         var st = stat()
         guard fstat(fd, &st) == 0, st.st_mode & S_IFMT == S_IFREG else { return nil }
         self.dev = Int64(st.st_dev)
-        self.ino = st.st_ino
+        self.ino = UInt64(st.st_ino)
     }
 
     /// Nil if `path` does not resolve to a regular file.
@@ -19,7 +19,7 @@ public struct HostFileIdentity: Equatable {
         var st = stat()
         guard stat(path, &st) == 0, st.st_mode & S_IFMT == S_IFREG else { return nil }
         self.dev = Int64(st.st_dev)
-        self.ino = st.st_ino
+        self.ino = UInt64(st.st_ino)
     }
 }
